fix(Map): simplify data preview by removing custom statistics management
Statistics are now shared with Geotk, therefore we do not need to prepare a style uphill.

This was dangerous, because at this point we don't get enough information about data structure.
Geotoolkit renderer has far more advanced logic, and will better handle styling as long as it can fetch statistics.