feat(Engine+Storage): Force SIS resource proxying to override metadata.
Examind save a copy of resource original metadata, to be able to add new information in it.
For now, this copy of the metadata is not accessible by low-level tools (Geotoolkit, SIS).
It causes problems, particularly to send image statistics to Geotoolkit renderer.

This commit aims to change that by proxying any resource registered.
It will force the resource to send back Examind metadata instead of its own.